Remove legacy webrtc provider APIs
